shiny new update
changes across the board, so lets call it a v2.99 update. not a v3
update because we are still using at the old susy grid. the plan is to
switch to a flex column layout and archive the grid. when that happens
that will make it officially a v3, imo.